+++ src/contract_wrappers/modules/module_factory_wrapper/common.ts
@@ -316,3 +316,7 @@ export default abstract class ModuleFactoryCommon extends ContractWrapper {
     );
   };
 }
+
+export function isModuleFactory(wrapper: ContractWrapper): wrapper is ModuleFactoryCommon {
+  return wrapper instanceof ModuleFactoryCommon;
+};