+++ src/contract_wrappers/modules/sto/capped_sto_wrapper/index.ts
@@ -1,12 +1,13 @@
 /* istanbul ignore file */
 import { CappedSTO_3_0_0, isCappedSTO_3_0_0 } from './3.0.0';
 import { CappedSTO_3_1_0, isCappedSTO_3_1_0 } from './3.1.0';
-import Common from './common';
+import Common, { isCappedSTO } from './common';
 import { ContractVersion, Subscribe, GetLogs } from '../../../../types';
 
 export type CappedSTO = CappedSTO_3_0_0 | CappedSTO_3_1_0;
 
 export {
+  isCappedSTO,
   CappedSTO_3_0_0,
   isCappedSTO_3_0_0,
   CappedSTO_3_1_0,