+++ src/contract_wrappers/modules/sto/usd_tiered_sto_wrapper/index.ts
@@ -1,12 +1,13 @@
 /* istanbul ignore file */
 import { USDTieredSTO_3_0_0, isUSDTieredSTO_3_0_0 } from './3.0.0';
 import { USDTieredSTO_3_1_0, isUSDTieredSTO_3_1_0 } from './3.1.0';
-import Common from './common';
+import Common, { isUSDTieredSTO } from './common';
 import { ContractVersion, Subscribe, GetLogs } from '../../../../types';
 
 export type USDTieredSTO = USDTieredSTO_3_0_0 | USDTieredSTO_3_1_0;
 
 export {
+  isUSDTieredSTO,
   USDTieredSTO_3_0_0,
   isUSDTieredSTO_3_0_0,
   USDTieredSTO_3_1_0,