+++ src/contract_wrappers/modules/transfer_manager/lock_up_transfer_manager_wrapper/index.ts
@@ -1,12 +1,16 @@
 /* istanbul ignore file */
 import { LockUpTransferManager_3_0_0, isLockUpTransferManager_3_0_0 } from './3.0.0';
 
-import Common from './common';
+import Common, { isLockUpTransferManager } from './common';
 import { ContractVersion, Subscribe, GetLogs } from '../../../../types';
 
 export type LockUpTransferManager = LockUpTransferManager_3_0_0;
 
-export { LockUpTransferManager_3_0_0, isLockUpTransferManager_3_0_0 };
+export {
+  isLockUpTransferManager,
+  LockUpTransferManager_3_0_0,
+  isLockUpTransferManager_3_0_0
+};
 
 // for internal use
 export class LockupTransferManagerCommon extends Common {