+++ src/contract_wrappers/modules/transfer_manager/percentage_transfer_manager_wrapper/index.ts
@@ -1,12 +1,16 @@
 /* istanbul ignore file */
 import { PercentageTransferManager_3_0_0, isPercentageTransferManager_3_0_0 } from './3.0.0';
 
-import Common from './common';
+import Common, { isPercentageTransferManager } from './common';
 import { ContractVersion, Subscribe, GetLogs } from '../../../../types';
 
 export type PercentageTransferManager = PercentageTransferManager_3_0_0;
 
-export { PercentageTransferManager_3_0_0, isPercentageTransferManager_3_0_0 };
+export {
+  isPercentageTransferManager,
+  PercentageTransferManager_3_0_0,
+  isPercentageTransferManager_3_0_0
+};
 
 // for internal use
 export class PercentageTransferManagerCommon extends Common {