+++ src/contract_wrappers/tokens/security_token_wrapper/index.ts
@@ -1,12 +1,16 @@
 /* istanbul ignore file */
 import { SecurityToken_3_0_0, isSecurityToken_3_0_0 } from './3.0.0';
 
-import Common from './common';
+import Common, { isSecurityToken } from './common';
 import { ContractVersion, Subscribe, GetLogs } from '../../../types';
 
 export type SecurityToken = SecurityToken_3_0_0;
 
-export { SecurityToken_3_0_0, isSecurityToken_3_0_0 };
+export {
+  isSecurityToken,
+  SecurityToken_3_0_0,
+  isSecurityToken_3_0_0
+};
 
 // for internal use
 export class SecurityTokenCommon extends Common {